Revert "Make notifications into proper middleware"
This reverts commit 5493a2af4e9c9f1c8b9b7f9d122bb8f4a1ce6440. As
discussed on Discord the factory actually is the best implementation.
